# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "McpPendingAssignmentsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/McpPendingAssignment"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "McpPendingAssignmentsResponse",
        "type": "object"
      },
...
      "McpSaveRubricRequest": {
        "additionalProperties": false,
        "description": "保存客户端提取的题目 rubric。\n\n``handle`` 为 ``open_ai_marking_assignment`` 返回的 rubric 提取句柄;\n``status`` 为 ``complete``（有明确 rubric）或 ``absent_or_ambiguous``\n（无明确 rubric）。``complete`` 时逐项 ``source_quote`` 必须是题目 OCR\n原文子串且包含该项满分,服务端确定性校验后写入题目级权威快照。",
        "properties": {
          "handle": {
            "maxLength": 2000,
            "minLength": 20,
            "title": "Handle",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/McpRubricExtractionItem"
            },
            "maxItems": 100,
            "title": "Items",
            "type": "array"
          },
          "status": {
            "enum": [
              "complete",
              "absent_or_ambiguous"
            ],
            "title": "Status",
            "type": "string"
          },
          "total_max_score": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Max Score"
          }
        },
        "required": [
          "handle",
          "status"
        ],
        "title": "McpSaveRubricRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "SURF-2026-0031 后端 API",
    "title": "AI 作业批改系统",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/mcp/pending-assignments": {
      "get": {
        "description": "列出等待 MCP 客户端评分的作业(按上传时间与 ID 升序)。\n\n仅列出 ``awaiting_mcp`` 的作业。使用 keyset 游标分页:\n``(uploaded_at, id) >`` 上一个游标,保证排序稳定且不随翻页偏移。\n``limit`` 范围 1–100。",
        "operationId": "mcp_pending_assignments_api_mcp_pending_assignments_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpPendingAssignmentsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mcp Pending Assignments",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/mcp/questions/{question_id}/rubric": {
      "put": {
        "description": "保存客户端从题目 OCR 提取的 rubric(服务端确定性校验)。\n\n- 校验 rubric 提取句柄有效、绑定 submission/question、OCR 未变化\n- ``complete`` 时逐项 source_quote 必须是 OCR 子串且含满分、分项=总分、\n  条目不重复,通过后写入题目级权威快照\n- ``absent_or_ambiguous`` 时持久化识别结果,后续评分走配置 rubric 或\n  内置默认,避免重复询问客户端\n- 保存后旧评分句柄失效,客户端必须重新 ``open_ai_marking_assignment``\n  以新 rubric 快照评分",
        "operationId": "save_mcp_question_rubric_api_mcp_questions__question_id__rubric_put",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpSaveRubricRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpSaveRubricResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Mcp Question Rubric",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/mcp/submission-preflight": {
      "post": {
        "description": "Resolve a teacher-provided question name and validate only file metadata.",
        "operationId": "mcp_submission_preflight_api_mcp_submission_preflight_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpPreflightRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpPreflightResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mcp Submission Preflight",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/mcp/submissions/{submission_id}/assessment-review": {
      "put": {
        "description": "保存独立复核任务对当前评分建议的复核结论。\n\n- 评分句柄必须来自完整读取 ``ready_for_review`` 作业的评分包\n- 句柄绑定的 context_hash 与 grading_revision 必须仍然匹配,\n  建议已更新时返回 409,复核者需重新打开作业\n- ``items`` 必须逐项引用当前 rubric 的全部 item(不重复、不缺失);\n  ``suggested_score`` 不得超过该项满分\n- 复核结论写入 ``submissions.assessment_review`` 供教师在网页参考,\n  不改变建议本身与状态;教师仍在网页确认最终成绩",
        "operationId": "save_mcp_assessment_review_api_mcp_submissions__submission_id__assessment_review_put",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "title": "Submission Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpSaveReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpSaveReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Mcp Assessment Review",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/mcp/submissions/{submission_id}/assessment-v2": {
      "put": {
        "description": "Save a model-facing compact assessment using a server-issued grading handle.",
        "operationId": "save_mcp_assessment_api_mcp_submissions__submission_id__assessment_v2_put",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "title": "Submission Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpSaveAssessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpAssessmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Mcp Assessment",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/mcp/submissions/{submission_id}/package": {
      "get": {
        "description": "Expose the grading context in bounded pages with opaque continuation.",
        "operationId": "mcp_submission_package_api_mcp_submissions__submission_id__package_get",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "title": "Submission Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "continuation_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Continuation Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpPackageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mcp Submission Package",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/mcp/submissions/{submission_id}/visual-confirmation": {
      "post": {
        "operationId": "confirm_mcp_visual_review_api_mcp_submissions__submission_id__visual_confirmation_post",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "title": "Submission Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpVisualConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpVisualConfirmationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Mcp Visual Review",
        "tags": [
          "mcp"
        ]
      }
    },
...
  }
}